+++ internal/services/gateway.go
@@ -58,14 +58,18 @@ func (svc *gatewayService) ForwardRequest(hst string, pthPfx string, schm string
 		// copy the inbound request to the proxy request state struct
 		ctx.Request.CopyTo(pxyReq.req)
 
-		// remap the host, path, and scheme to downstream server
+		// evaluate inbound path and determine if adjustments are needed
 		dwnUri := pxyReq.req.URI()
-
 		pth := string(dwnUri.Path())
-		if pthPfx != "/" {
+		if len(pthPfx) > 1 {
+			svc.log.Trace().
+				Str("prefix", pthPfx).
+				Str("path", pth).
+				Msg("Trimming prefix from path")
 			pth = strings.TrimPrefix(pth, pthPfx)
 		}
 
+		// 
 		svc.log.Debug().
 			Str("originalHost", string(dwnUri.Host())).
 			Str("originalPath", string(dwnUri.Path())).